프로파일링 마커 추가 및 렌더러 GPU에셋 업데이트 시기 수정

+++ SSGAL/SSGAL/Private/DX12/GALResourceUpdater/DX12GALResourceUpdater.h
@@ -33,7 +33,7 @@ public:
 		D3D12_RESOURCE_STATES ToState);
 
 private:
-	ID3D12Resource* RentUpdateBuffer(int32& OutBufferStartOffset, int32 BufferSize);
+	ID3D12Resource* RentUpdateBuffer(int32& OutBufferStartOffset, int64 BufferSize);
 
 private:
 	SS::PooledList<ID3D12Resource*> _DefaultUploadBuffers;
